refactor(security): enforce immutability and improve token handling
- Made all entity fields immutable to enhance consistency and thread safety.
- Changed `userId` type from `Long` to `Integer` in `UserPrincipal` and `TokenClaimsDto` for alignment with the user model.
- Refactored `saveUserToken` method to use the builder pattern for clearer and safer object construction.

+++ services/bank-security/src/main/java/am/banking/system/security/model/dto/UserPrincipal.java
@@ -21,14 +21,14 @@ public class UserPrincipal implements UserDetails, Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private final Long userId;
+    private final Integer userId;
     private final String username;
     private final String password;
     private final String email;
     private final Set<String> roles;
     private final Set<String> permissions;
 
-    public UserPrincipal(Long userId, String username, String password, String email, Set<String> roles, Set<String> permissions) {
+    public UserPrincipal(Integer userId, String username, String password, String email, Set<String> roles, Set<String> permissions) {
         this.userId = userId;
         this.username = username;
         this.password = password;
