+++ services/bank-security/src/main/java/am/banking/system/security/token/claims/TokenClaimsDto.java
@@ -13,6 +13,6 @@ import java.util.Set;
  * Time: 01:36:08
  */
 @Builder
-public record TokenClaimsDto(Long userId, String username, String email, Set<String> roles, Set<String> permissions,
+public record TokenClaimsDto(Integer userId, String username, String email, Set<String> roles, Set<String> permissions,
                              TokenState tokenState, TokenPurpose tokenPurpose, AccountState accountState) {
 }
